TECH DEBT - Constrain the length of UKPRNs to 8 digits (int)

+++ src/SFA.DAS.ApplyService.InternalApi/Infrastructure/IUkrlpSoapSerializer.cs
@@ -4,7 +4,7 @@ namespace SFA.DAS.ApplyService.InternalApi.Infrastructure
 {
     public interface IUkrlpSoapSerializer
     {
-        string BuildUkrlpSoapRequest(long ukprn, string stakeholderId, string queryId);
+        string BuildUkrlpSoapRequest(int ukprn, string stakeholderId, string queryId);
         MatchingProviderRecords DeserialiseMatchingProviderRecordsResponse(string soapXml);
     }
 }
